Add checking for related route stations before deleting station row.

+++ app/src/main/java/com/example/coursework/ui/station/viewModels/StationDetailsViewModel.kt
@@ -4,19 +4,24 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.coursework.data.repositories.StationsRepository
+import com.example.coursework.data.repositories.relationsRepositories.StationWithRouteStationsRepository
 import com.example.coursework.ui.state.StationUiState
 import com.example.coursework.ui.state.toStation
 import com.example.coursework.ui.state.toStationUiState
 import com.example.coursework.ui.station.screens.StationDetailsDestination
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.*
 
 
 class StationDetailsViewModel(
     savedStateHandle: SavedStateHandle,
-    private val stationsRepository: StationsRepository
+    private val stationsRepository: StationsRepository,
+    private val stationWithRouteStationsRepository: StationWithRouteStationsRepository
 ) : ViewModel() {
 
-    private val stationId: Int = checkNotNull(savedStateHandle[StationDetailsDestination.stationIdArg])
+    private val stationId: Int =
+        checkNotNull(savedStateHandle[StationDetailsDestination.stationIdArg])
     val uiState: StateFlow<StationUiState> = stationsRepository.getStationStream(stationId)
         .filterNotNull()
         .map {
@@ -27,8 +32,17 @@ class StationDetailsViewModel(
             initialValue = StationUiState()
         )
 
-    suspend fun deleteItem() {
-        stationsRepository.deleteStation(uiState.value.toStation())
+    suspend fun validateStationDeletion(): String {
+        val message = viewModelScope.async(Dispatchers.IO) {
+            val stationWithRouteStationsList = stationWithRouteStationsRepository.getStationAndRouteStations()
+            if (stationWithRouteStationsList.any { uiState.value.id == it.station.id && it.routeStations.isNotEmpty() }) {
+                "This station can't be deleted because it's used in existing route station(-s)."
+            } else {
+                stationsRepository.deleteStation(uiState.value.toStation())
+                "Row deleted successfully."
+            }
+        }
+        return message.await()
     }
 
     companion object {
